finish renaming of data-type including filename
