+++ src/app/core/entity/schema/entity-schema.service.ts
@@ -32,7 +32,7 @@ import { mapEntitySchemaDatatype } from "../schema-datatypes/datatype-map";
 import { booleanEntitySchemaDatatype } from "../schema-datatypes/datatype-boolean";
 import { entityArrayEntitySchemaDatatype } from "../schema-datatypes/datatype-entity-array";
 import { entityEntitySchemaDatatype } from "../schema-datatypes/datatype-entity";
-import { dateWithAgeEntitySchemaDatatype } from "../schema-datatypes/datatype-date-of-birth";
+import { dateWithAgeEntitySchemaDatatype } from "../schema-datatypes/datatype-date-with-age";
 
 /**
  * Transform between entity instances and database objects
